Update DecryptFilterPlugin.java
Update from previous commit to edit the file in the project instead of a adding a new file. 
Changes to support prepended IV byte data with a new optional iv_block_size parameter for anyone else that uses this type of implementation for decryption. This change would be helpful so I do not have to use a custom implementation of this plugin